Tutorial_3_RabbitMQ_Docker/writer/requirements.txt
pika==1.1.0
flask==1.1.2
slackclient==2.9.3